Add a Landing page
Added a landing page to the blog so that users are
greeted with some text when they go to the
homepage. I also added a "toolbar-buffer" because
the toolbar is positioned absolutely on the page
so toolbar buffer allows all the rest of the
content on the page to be displayed correctly
below the toolbar. Only other major change is that
I updated the link in the toolbar to point to the
GitHub repo that contains this project.