Add generic Content Types
Added three generic content types for displaying blog posts: Code, Text, and
Image. There may be more added down the line, but for now, These will be a
good starting point. Implementation of this, along with the type checking
system was followed pretty much wholesale from this blog post, may be
interesting to write my own blog post about it. Next time we'll make
a component to display the blog posts more beautifully on the homepage.

Link: https://spin.atomicobject.com/2021/12/01/ngswitch-type-errors/