Add First Post
This is the code allowed me to make my first post. It's now up on firebase
and it gets read correctly on the homepage. Up next will be to update
the post component so that it actually reads the post data and displays it
instead of the dummy data I've got in there now. Then it'll start to be
a real blog! I also updated the database to use numeric timestamps to
store the times that posts get created.