Add Database call to Post Display
Changed the Post Component to display data from the database instead of
the stock post I had originally. Now that I can create posts and view
posts from database, the basic functionality of the blog is complete.
Next, we'll update the look of the blog, add the ability to link one
blog post to the next, and verify that they get sorted correctly on the
homepage. Part of doing that will be adding more blog posts.