Update Post Display
Updated the CSS in the post component, to make the code look better and
the font for the paragraphs bigger. I also fixed an issue with
submitting a post, I had forgotten to add checks for the newly add
Heading and Link content types, in the CreateContent function call.
I might be good to think about how to solve that in the future, so I
don't screw up the database calls when I add content types. Made my
second blog post today and it looks much better than the first.
Hopefully next time we can update Firebase permissions so that I can be
the only one who submits blog posts.