Add Route Protection and Google Login
Added a Route Guard to the create post page to prevent people who are
not logged in as joshbarthelmess3@gmail.com from accessing the page.
This will prevent some of attempts to maliciously add to the database,
but next will be to update the permissions on the database itself. I
don't imagine this will be represented as a large code change, but I'll
at least make a blog entry about it.