Add Minesweeper Game
This is a really big commit, but I'll start with the small stuff. I
removed the toolbar from the app component page and made it it's own
component so that it could be added or removed from pages as needed. I
then added it to the pages it was already on (post, create-post, and
homepage).

The bigger item is the minesweeper world folder which contains the code
used to play the game I'm calling minesweeper world. It's a version of
minesweeper which wraps around on itself, such that it has no edges. The
board is big enough such that it overflows the screen and allows the
player to drag the world around to get to different areas of the map and
keep playing. Completing a game will be a large undertaking. I was going
to use the Angular DragAndDrop Module, but after determining that I needed
too much custom behavior, I implemented the feature myself. The basic
mechanics of the game are complete, but there are still a few things to
add, like a modal for instructions, a pop up to show when you win or
lose, a timer, to show how long you've been playing, and a link back to
the homepage.

Things to talk about in the blog posts about this: implementing drag and
drop, adding boolean checks to prevent event listeners from firing,
solving drift with the top and left anchor, the difference between
absolute and fixed positioning, working around the lack of capturing
events in Angular, and the actual mechanics of minesweeper.