Add Syntax Highlighting
Added Prism.js to allow our code blocks to show code with syntax
highlighting. We just use the standard theme for JavaScript, we might
mess with it later. In the future, we'll want to add an option on the
add post page to indicate what type of code is in the code block. We
also refactored the getPostContent function from the firebase service to
be an async function that no longer returns a post that hasn't loaded
its content yet. Might be good to talk about in a blog post.

Closes #9 #10 #13 #14