Added a Basic Pokemon List Game
Added a game to the site where you can try to list all the pokemon from
a given generation within a set time limit. It's not pretty and needs a
win and lose screen along with instructions, and support for generations
5-8 but the basic functionality works. Also want to add basic pokemon data
pages that the player can view by clicking on the pokemon in the list.