Finished List-em-all Game Functionality
Added instructions, a win and lose screen, and the last 4 generations.
The list also starts at the correct number given the generation. For now
that should be enough, it would be cool to be able to click on pokemon
to see info about them. Need to add a link to it on the home screen, a
link back to the homescreen.