Updated Links On Homepage
I updated the links to the games on the homepage and added the toolbar
at the top of the list em all game to ensure that player could easily
navigate back to the homescreen. I also made some minor tweaks to the
list em all game to improve readability, and overall style. I added the
red highlighting on the pokemon that were missed, and added a "please
select generation" message when the player hasn't yet chosen a generation
to play on.